+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Revision history for bejnarkli
 
+## Unreleased
+
+* Add --listenaddress option
+
 ## 0.2.0.1 -- 2021-01-05
 
 * Fix a circular dependency in systemd unit ordering.